+ strongruby23.y: add untyped variants to spec.
For each supported flavor of type annotation, allow the regular untyped
syntax to coexist. Typed variants are given by new *ty nodes with an
additional expression node that will contain the type, recognized as an
arbitrary expression. These corresponding positions will be handled by
the `primary` parser reduction rule.

As an alternative, StrongRuby would always generate typed nodes, with or
without the last expression depending on the presence of a type
annotation.