+ strongruby23.y: simple return type annotations.
Add support for return type annotations to the StrongRuby grammar. In
order to keep the grammar simple, it is sufficient to allow a
simplification of the regular type annotation syntax, consisting of
`tCONSTANT` instead of `primary` expressions for the types.

It must be noted that `tCOLON` type annotations may compete to bind to
either the last unparenthesized argument or the return value of the
method, in which case the argument takes precedence.